pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blink-stitch"
version = "0.0.0"
description = "Blink multicamera stitching utilities (packaged)"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9,<=3.12"
authors = [
  { name = "Dr Andrew M. Cox", email = "dev@andrewcox.doctor" }
]
dependencies = [
  "faster_whisper",
  "hdbscan",
  "humanize",
  "inaspeechsegmenter",
  "librosa",
  "loguru",
  "numpy",
  "opensmile",
  "pandas",
  "psutil",
  "pyannote.audio",
  "pyannote.core",
  "pyyaml",
  "rich",
  "scikit-learn",
  "torch",
  "torchaudio<2.8.0",
]

[project.urls]
Homepage = "https://example.com"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
blink-stitch = "blink_stitch.cli:main"
